moved `tags` & `self_roles` into the `staff_commands` package
