+++ src/main/java/net/javadiscord/javabot/Bot.java
@@ -28,14 +28,14 @@ import net.javadiscord.javabot.systems.moderation.report.ReportManager;
 import net.javadiscord.javabot.systems.moderation.server_lock.ServerLockManager;
 import net.javadiscord.javabot.systems.qotw.commands.questions_queue.AddQuestionSubcommand;
 import net.javadiscord.javabot.systems.qotw.submissions.SubmissionInteractionManager;
-import net.javadiscord.javabot.systems.self_roles.SelfRoleInteractionManager;
+import net.javadiscord.javabot.systems.staff_commands.self_roles.SelfRoleInteractionManager;
 import net.javadiscord.javabot.systems.staff_commands.embeds.AddEmbedFieldSubcommand;
 import net.javadiscord.javabot.systems.staff_commands.embeds.CreateEmbedSubcommand;
 import net.javadiscord.javabot.systems.staff_commands.embeds.EditEmbedSubcommand;
 import net.javadiscord.javabot.systems.starboard.StarboardManager;
-import net.javadiscord.javabot.systems.tags.CustomTagManager;
-import net.javadiscord.javabot.systems.tags.commands.CreateCustomTagSubcommand;
-import net.javadiscord.javabot.systems.tags.commands.EditCustomTagSubcommand;
+import net.javadiscord.javabot.systems.staff_commands.tags.CustomTagManager;
+import net.javadiscord.javabot.systems.staff_commands.tags.commands.CreateCustomTagSubcommand;
+import net.javadiscord.javabot.systems.staff_commands.tags.commands.EditCustomTagSubcommand;
 import net.javadiscord.javabot.systems.user_commands.leaderboard.ExperienceLeaderboardSubcommand;
 import net.javadiscord.javabot.tasks.MetricsUpdater;
 import net.javadiscord.javabot.tasks.PresenceUpdater;
