+++ src/main/java/net/javadiscord/javabot/systems/staff_commands/self_roles/SelfRoleCommand.java
@@ -1,4 +1,4 @@
-package net.javadiscord.javabot.systems.self_roles;
+package net.javadiscord.javabot.systems.staff_commands.self_roles;
 
 import com.dynxsty.dih4jda.interactions.commands.SlashCommand;
 import net.dv8tion.jda.api.interactions.commands.DefaultMemberPermissions;
